+++ CHANGELOG.md
@@ -13,3 +13,14 @@ All notable changes to this project will be documented in this file.
 - Logging configuration
 - Health probes
 - Prometheus metrics
+
+## [V0.0.2](https://github.com/hary-singh/spring-boot-commons/releases/tag/v0.0.2) - 2025-05-25
+
+### Added
+
+- Support for Spring Boot 3.4.5
+- Improved OpenTelemetry service name resolution
+
+### Fixed
+
+- Improved service name resolution in OpenTelemetryConfig to prioritize OTEL_SERVICE_NAME environment variable